# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CommissionResponse:
    properties:
      commission_percentage:
        type: number
      id:
        type: string
      income_amount:
        type: number
      is_active:
        type: boolean
      level:
        type: integer
      seat_capacity:
        type: integer
    type: object
...
  models.Income:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      from_member:
        $ref: '#/definitions/models.Member'
      from_member_id:
        type: string
      id:
        type: string
      level:
        type: integer
      member:
        allOf:
        - $ref: '#/definitions/models.Member'
        description: Relationships
      member_id:
        type: string
      percentage:
        type: number
      processed_at:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
  models.Member:
    properties:
      created_at:
        type: string
      downlines:
        description: Relationships
        items:
          $ref: '#/definitions/models.Member'
        type: array
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      incomes:
        items:
          $ref: '#/definitions/models.Income'
        type: array
      joined_at:
        type: string
      member_code:
        type: string
      phone:
        type: string
      referrals:
        items:
          $ref: '#/definitions/models.Referral'
        type: array
      sponsor:
        $ref: '#/definitions/models.Member'
      sponsor_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.MemberResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      joined_at:
        type: string
      member_code:
        type: string
      phone:
        type: string
      sponsor_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  models.MemberWithLevel:
    properties:
      created_at:
        type: string
      downlines:
        description: Relationships
        items:
          $ref: '#/definitions/models.Member'
        type: array
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      incomes:
        items:
          $ref: '#/definitions/models.Income'
        type: array
      joined_at:
        type: string
      member_code:
        type: string
      phone:
        type: string
      referrals:
        items:
          $ref: '#/definitions/models.Referral'
        type: array
      relationship_level:
        type: integer
      sponsor:
        $ref: '#/definitions/models.Member'
      sponsor_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.Referral:
    properties:
      child:
        $ref: '#/definitions/models.Member'
      child_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      level:
        type: integer
      parent:
        allOf:
        - $ref: '#/definitions/models.Member'
        description: Relationships
      parent_id:
        type: string
    type: object
...
  services.GrowthProjection:
    properties:
      description:
        type: string
      percentage:
        type: number
      total_income:
        type: number
    type: object
  services.GrowthProjectionResponse:
    properties:
      level:
        type: integer
      projections:
        items:
          $ref: '#/definitions/services.GrowthProjection'
        type: array
    type: object
...
  services.IncomeProjectionResponse:
    properties:
      growth_projections:
        items:
          $ref: '#/definitions/services.GrowthProjection'
        type: array
      level_projections:
        items:
          $ref: '#/definitions/services.LevelProjection'
        type: array
      member_id:
        type: string
      total_actual:
        type: number
      total_potential:
        type: number
    type: object
  services.LevelBreakdown:
    properties:
      income_amount:
        type: number
      level:
        type: integer
      percentage:
        type: number
      seat_capacity:
        type: integer
      seat_filled:
        type: integer
      total_income:
        type: number
    type: object
  services.LevelDetail:
    properties:
      actual_income:
        type: number
      income_per_seat:
        type: number
      level:
        type: integer
      members:
        items:
          $ref: '#/definitions/models.MemberWithLevel'
        type: array
      potential_income:
        type: number
      seat_capacity:
        type: integer
      total_members:
        type: integer
    type: object
...
  services.LevelProjection:
    properties:
      actual_income:
        type: number
      income_per_seat:
        type: number
      level:
        type: integer
      percentage_complete:
        type: number
      potential_income:
        type: number
      seat_capacity:
        type: integer
      seat_filled:
        type: integer
    type: object
...
  services.ReferralStatsResponse:
    properties:
      direct_referrals:
        type: integer
      indirect_referrals:
        type: integer
      level_distribution:
        additionalProperties:
          type: integer
        type: object
      max_tree_depth:
        type: integer
      total_referrals:
        type: integer
    type: object
  services.ReferralTreeResponse:
    properties:
      levels:
        items:
          $ref: '#/definitions/services.LevelDetail'
        type: array
      max_depth:
        type: integer
      member:
        $ref: '#/definitions/models.MemberResponse'
      total_downline:
        type: integer
    type: object
...
  services.TreeSummaryResponse:
    properties:
      active_members:
        type: integer
      level_breakdown:
        items:
          $ref: '#/definitions/services.LevelBreakdown'
        type: array
      potential_income:
        type: number
      total_income:
        type: number
      total_levels:
        type: integer
      total_members:
        type: integer
    type: object
  services.UpdateCommissionInput:
    properties:
      commission_percentage:
        type: number
      income_amount:
        type: number
      seat_capacity:
        type: integer
    type: object
  utils.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    Admin API for Rudra management system

    This API provides endpoints for managing Rudra operations including:
    - Admin authentication and authorization
    - Member management
    - Referral tree management
    - Income calculations and distributions
    - Reports and analytics
    - Audit logging
  title: Rudra Admin API
  version: 1.0.0
paths:
...
  /api/v1/referrals/{id}/downline:
    get:
      description: Get the downline referral tree for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Max depth level
        in: query
        name: level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.ReferralTreeResponse'
              type: object
      security:
      - Bearer: []
      summary: Get downline tree
      tags:
      - referrals
  /api/v1/referrals/{id}/income-projection:
    get:
      description: Get income projections for a member across all levels
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.IncomeProjectionResponse'
              type: object
      security:
      - Bearer: []
      summary: Get income projection
      tags:
      - referrals
  /api/v1/referrals/{id}/stats:
    get:
      description: Get referral statistics for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.ReferralStatsResponse'
              type: object
      security:
      - Bearer: []
      summary: Get referral stats
      tags:
      - referrals
  /api/v1/referrals/{id}/summary:
    get:
      description: Get a summary of the referral tree for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.TreeSummaryResponse'
              type: object
      security:
      - Bearer: []
      summary: Get tree summary
      tags:
      - referrals
  /api/v1/referrals/{id}/upline:
    get:
      description: Get the upline chain for a member
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Max depth level
        in: query
        name: level
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.ReferralTreeResponse'
              type: object
      security:
      - Bearer: []
      summary: Get upline
      tags:
      - referrals
  /api/v1/referrals/commission-config:
    get:
      description: Get the Rudra commission structure for all levels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CommissionResponse'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get commission config
      tags:
      - referrals
  /api/v1/referrals/commission-config/{level}:
    put:
      consumes:
      - application/json
      description: Update commission configuration for a specific level
      parameters:
      - description: Level number
        in: path
        name: level
        required: true
        type: integer
      - description: Commission config update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateCommissionInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.CommissionResponse'
              type: object
      security:
      - Bearer: []
      summary: Update commission config
      tags:
      - referrals
  /api/v1/referrals/projected-growth/{level}:
    get:
      description: Calculate projected income growth at different percentages for
        a level
      parameters:
      - description: Level number (1-10)
        in: path
        name: level
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/services.GrowthProjectionResponse'
              type: object
      security:
      - Bearer: []
      summary: Calculate projected growth
      tags:
      - referrals
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"